refactors waypoints
