+++ Assets/Scripts/EnemyWaypoints.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cc9088f511bf1c345a50adea397c81d5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
