+++ Assets/Scripts/EnemyWaypoints/WaypointsPath.cs.meta
